+++ HISTORY.md
@@ -1,4 +1,10 @@
 
+0.0.3 / 2011-08-05 
+==================
+
+  * Added support for LiveWhale Places
+  * Fixed naming bugs in /live/ REST api responders
+
 0.0.1 / 2011-06-14 
 ==================